Fixed reupload issue and provide db migrations

+++ Tinifier.Core/Models/API/TinifyImageStatistic.cs
@@ -15,6 +15,6 @@
         /// <summary>
         /// Total saved bytes for user
         /// </summary>
-        public int TotalSavedBytes { get; set; }
+        public long TotalSavedBytes { get; set; }
     }
 }
